Refactor Bottom Nav Bar to views/widgets
